requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.4
django-appconf==1.0.6
django-compressor==4.5.1
django-libsass==0.9
dnspython==2.7.0
et_xmlfile==2.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
libsass==0.23.0
mysql-connector-python==9.2.0
numpy==2.2.2
openpyxl==3.1.5
pandas==2.2.3
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
rcssmin==1.1.2
regex==2024.11.6
requests==2.32.3
rjsmin==1.2.2
rsa==4.9
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
gunicorn==20.1.0
matplotlib
seaborn